step 03: start real implementation (#3)
Basically rewrote it all. Cover the files in order that they are c.shalled.

docker/Dockerfile was updated to use the ISCE2 base image.

run_pge.sh is the entry point. This is mainly needed to setup the ISCE environment as well as activate conda.

iterate.py was cleaned up and updated to reflect state information that needs to travel with the AOI.

active.py is responsible for processing an AOI track that is still active. Active is defined as the AOI track endtime being greater than now with now meaning the time of execution. It starts with active.process() and branches throughout. I fills the AOI with older data over the area then starts working on time after the event. The after event loop is written to process data that is already there until present. This allows historical events to be declared and they will generate all of their acquisition lists immediately.

es is the elastic search package. It hides the query and request details from the rest of the code. It uses the http interface because that is the example used as a template. It could be replaced and none of the rest of the code would notice.

footprint.py computes and handles all of the acquisition to footprint conversions and computations.

orbit.py deals with orbit files (EOFs). It loads them and converts them to ISCE objects. Again it is a helper module to hide much of the gory details.

slc.py hides parsing the acquisitions into acquisition lists for the localizer. It iterates through all of the current acquisitions (primary or reference) and matches to past acquisitions (secondary). It then sends that grouping off as an acquisition list to the localizer.

test.py is a helper module to run some sanity tests.