step 04: parameterization (#4)
* fixed up docker/hysds-io.json

Removed the old elements within the file that are not used. Then added the correct parameters, types, and casting to the file.

* fixed up docker/job-spec.json

Altered its entries to reflect those in hysds-io.json

* added context.py

This is a helper module that hides the reading and access to the _context.json file used in the container at run time. Instead it provides a uniform interface to hide where and how it is actually running.