Step 05: code and name cleanup (#5)
* moved to coseismic_enumerator

Not sure where co-event came from but rest of project seems to have moved to coseismic so joined it. Also, this work is just an enumerator and no has a submitter element so that is how the new name was arrived at.

* removed some old files

Used files for cribbing information from but now that the job runs removed them.