Step 07: hackathon to integrate (#8)
* fixed the segfault

Turns out there are many pythons running in the docker container and they do not play well together in all permutations. This commit records a working permutation that does work for both isce and hysds tools to be used together.

* fixes

footprint.convert() updated to generate a correct json object that osgeo can gutn into a Geometry. Found this problem and fix when it kept returning a None for the create.

footprint.coverage() updated because osgeo kept complaining that two areas did not overlap and thus no union. Bad process was to union all of the footprints into a mega area then intersect it with the AOI then compute that intersections area. Because some footprints may be detached and not have a union, intersect each bit with the AOI and compute its area which logically can be 0. Then sum them all for total coverage. Might be a little weird if they all cover the same place that can then tally more than the whole.

footprint.intersection_area() is a new function that computes the area of intersection between the AOI and a footprint.

footprint.project() updated to change from a zip object to an actual list. json.dumps() does not do this for us which causes problems.

footprint.union() deleted.

orbit.load() did not compute the local filename correctly. Fixed.

slc.load() updated to fix intersection problem. osgeo returns a None when there is no intersection rather than an object that has an Ares() of 0. Put in the additional checking.

* correct coverage

footprint.coverage() summed all of the independent footprints. This does not account for the overlap of the footprints with themselves. Changed the algorithm to compute all of the largest independent (no overlap) footprints then intersect those with the AOI for an accurate measure of the coverage.

footprint.unionize() merges a list of footprints into the smallest set of independent (no intersection or overlap) footprints of largest size. It is recurisive and the python construct for/else is somewhat tricky because many do not use it.

* handle broken EOF files

At least one of the orbit files is incomplete causing coverage() to fail. Changed it such that a failure to read the EOF forces coverage() to return 0. In turn, the algorithm will simply move backward or forward in time to find another set of acquisitions and orbits to use.

* add clean up

It looks like hysds is trying to ingest the orbit files that were downloaded. Removing these at the end of each AOI to conserve diskspace as orbit files should be limited in scope and therefore not be overlapping all that much.