step 08: prepare for last hackathon (#11)
* run single and queues

Updated docker/hysds-io JSON to be individual instead of iteration. When jobs are run, elastic search is used to parallel run the jobs with a single job for each item found. Hence the term iteration. Setting it to individual ignore elastic search and starts just a single job.

Updated docker/job-io JSON with the new queue.

* added filters for testing

Two filters where added to activate.py to constrain how much data was processed decreasing the processing time and thus turn around time.

The first filter was to not process zesty-test-data* and should be removed when convinced ready to update AOI-tracks for real. 

The second filter was to do a specific area of interest only. This too should be removed when all testing is done.

* more metadata

Need to add more and more metadata to the acq list. Updated slc.py to do this and as a consequence need to pass the aoi from activate.py to slc.py.

* general output improvement

README.md, es/__init__.py, run_pge.sh, all received some improvements.