Issue 15: further constraints for topsapp (#16)
* constrain by track number

Too many crossings were being detected in the span of hours to days that are mostly not usable. Added the track number constraint to only find those crossings with the same track number as defined in the earthquake AOI metadata.

* reshape the acquisition list generation (slc.load())

The looping needed to become more complicated so that any clumped acquisitions are formed to a single acquisition list. This in combination with track number constraint should make it one reference (master) and one or two secondary (slaves). It can be two if the data take stopped and restarted within the AOI.  It is a somewhat pathological case but is possible.



* new constants

As part of the SLC reshape, needed to move some the constants to a new module to prevent cyclic or circular importing.