enumeration updates
* active.py

Add two behaviors. First, footprint.prune() removes those acquisitions that have a less than a threshold of overlap with the AOI. This seems to be happening on the boundaries of the AOI where an acquisition ends on that boundary and the next acquisition has just a touch of overlap as all acquisitions do. The threshold is determined direction from the coverage_threshold_percent by using its inverse (100 - coverage_threshold_percent). Since that value is currently set at 90%, then overlaps of less than 10% are discarded. If it is set to 95% then less than 5% is discarded. The new behavior is the use of the track number as a filter parameter when using elastic search to locate all acquisition that intersect with the AOI.

* context.py

Added new value that the launcher of the PGE: reset_all. reset_all is false when 0 and true other wise. See iterate.py for the behavioral changes in the system.

* footprint.py

Added prune(). This is a particularly evil function as it modifies the lists it is given which means it has side effects which are evil. Side effects were chosen because it was the simplest and least expensive way to insert the functionality. Basically, it walks through the acquisition and orbit lists and removes the pairs where the acquisition does not have enough overlap with the AOI. The threshold for enough overlap is 100 - coverage_threshold_percent.

* iterate.py

Added a new behavior that allows AOIs that have already been processed to be treated as new. The reset could be made more specific, but it is currently the simplest of implementations allow all or none of the AOIs to be processed as new independent of them being processed previously. Use with care.

* slc.py

This module had the most changes but the least change in behavior. There are 3 ways to generate acquisitions lists that we as a team have been trying out. Simply coded up all three versions and then use an if statement in load() to determine which algorithm to use.