change from location to water masked location
Put a commonly usable function to extract the appropriate location from a geolocated object. If it finds the water masked location it will use that but will fall back to the general location when not present.

All uses of [location] are changed to get_location i active.py and footprint.puy